requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
decorator==4.4.2
fastapi==0.101.1
ffmpeg==1.4
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
gradio_client==0.5.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
Jinja2==3.1.2
MarkupSafe==2.1.3
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.25.2
openai==0.27.9
packaging==23.1
Pillow==10.0.0
playwright==1.37.0
proglog==0.1.10
pydantic==2.3.0
pydantic_core==2.6.3
pyee==9.0.4
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
replicate==0.11.0
requests==2.31.0
safetensors==0.3.3
scipy==1.11.2
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.1
transformers==4.32.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
websockets==11.0.3
yarl==1.9.2
